requirements.txt
python-dotenv~=1.0.1
openai~=1.58.1
schedule~=1.2.2
streamlit~=1.41.1
selenium~=4.27.1
requests~=2.32.3
python-dateutil~=2.9.0.post0
webdriver-manager~=4.0.2